requirements.txt
pandas>=1.5.0
numpy>=1.23.0,<2.0.0
scikit-learn>=1.2.0
scipy>=1.10.0
scikit-surprise>=1.1.3
streamlit>=1.32.0
plotly>=5.18.0
